requirements.txt
typer>=0.9.0
rich>=13.0.0
prompt-toolkit>=3.0.38
langchain>=0.2.0
langchain-community>=0.2.0
langchain-ollama>=0.1.0
langchain-core>=0.2.0
langchain-openai>=0.1.0
langchain-google-genai>=0.1.0
ollama>=0.2.0
pydantic>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0
llama-cpp-python>=0.2.75
huggingface_hub>=0.23.0
pyfiglet>=1.0.2
